Fix an exception-on-quit in window-closing callback
If there was a saved session, the login window is never shown, and there is
never an auth token. When the app quits, the web view window "closes", and the
closing callback runs. Because there is no auth token, it tries to insert nil
into a dictionary, which causes an exception.

A nil check on authToken is not enough, because there will be another use for
the web view window in the near future (getting the read_notifications extended
permission).

The state-machine construction is nasty, but the state has to be stored
somewhere.